moving decoding to FeedItems designated FeedItemsMapper class
