+++ FeedAPIChallenge/Feed API/RemoteFeedLoader.swift
@@ -19,49 +19,22 @@ public final class RemoteFeedLoader: FeedLoader {
 	}
 	
 	public func load(completion: @escaping (FeedLoader.Result) -> Void) {
-		client.get(from: url) { [unowned self] result in
+		client.get(from: url) { result in
 			switch result {
 			
 			case .failure:
 				completion(.failure(Error.connectivity))
 				
 			case .success(let(data, response)):
-				
-				guard response.statusCode == 200,
-					  let root = rootObject(from: data) else {
-					completion(.failure(Error.invalidData))
-					return
+
+				do {
+					let images = try FeedItemsMapper.map(data, response: response)
+					completion(.success(images))
 				}
-				
-				completion(.success(root.images))
+				catch {
+					completion(.failure(error))
+				}
 			}
-		}
-	}
-	
-	private func rootObject(from data: Data) -> Root? {
-		try? JSONDecoder().decode(Root.self, from: data)
-	}
-	
-	private struct Root: Decodable {
-		
-		private struct FeedItem: Decodable {
-			let id: UUID
-			let description: String?
-			let location: String?
-			let url: URL
-			
-			private enum CodingKeys: String, CodingKey {
-				case id = "image_id"
-				case description = "image_desc"
-				case location = "image_loc"
-				case url = "image_url"
-			}
-		}
-		
-		private let items: [FeedItem]
-		
-		var images: [FeedImage] {
-			items.map { FeedImage(id: $0.id, description: $0.description, location: $0.location, url: $0.url) }
 		}
 	}
 }
